Move the VQ-VAE trainer class to modules file
