+++ recognition/VQ-VAE-46495408/dataset.py
@@ -1,5 +1,4 @@
 import tensorflow as tf
-#import tensorflow_probability as tfp
 import matplotlib.pyplot as plt
 
 def get_train_dataset():
