+++ recognition/VQ-VAE-46495408/modules.py
@@ -82,3 +82,50 @@ def get_vqvae(latent_dim=16, num_embeddings=64):
     quantized_latents = vq_layer(encoder_outputs)
     reconstructions = decoder(quantized_latents)
     return keras.Model(inputs, reconstructions, name="vq_vae")
+
+class VQVAETrainer(keras.models.Model):
+    def __init__(self, train_variance, latent_dim=32, num_embeddings=128, **kwargs):
+        super().__init__(**kwargs)
+        self.train_variance = train_variance
+        self.latent_dim = latent_dim
+        self.num_embeddings = num_embeddings
+        
+        self.vqvae = get_vqvae(self.latent_dim, self.num_embeddings)
+        
+        self.total_loss_tracker = keras.metrics.Mean(name="total_loss")
+        self.reconstruction_loss_tracker = keras.metrics.Mean(name="reconstruction_loss")
+        self.vq_loss_tracker = keras.metrics.Mean(name="vq_loss")
+    
+    @property
+    def metrics(self):
+        return [
+            self.total_loss_tracker,
+            self.reconstruction_loss_tracker,
+            self.vq_loss_tracker,
+        ]
+
+    def train_step(self, x):
+        with tf.GradientTape() as tape:
+            # Outputs from the VQ-VAE
+            reconstructions = self.vqvae(x)
+            # Calculates the losses
+            reconstructions_loss = (
+                tf.reduce_mean((x - reconstructions) ** 2) / self.train_variance
+            )
+            total_loss = reconstructions_loss + sum(self.vqvae.losses)
+            
+        # Backpropagation
+        grads = tape.gradient(total_loss, self.vqvae.trainable_variables)
+        self.optimizer.apply_gradients(zip(grads, self.vqvae.trainable_variables))
+        
+        # Loss tracking
+        self.total_loss_tracker.update_state(total_loss)
+        self.reconstruction_loss_tracker.update_state(reconstructions_loss)
+        self.vq_loss_tracker.update_state(sum(self.vqvae.losses))
+        
+        # Log results
+        return {
+            "loss": self.total_loss_tracker.result(),
+            "reconstruction_loss": self.reconstruction_loss_tracker.result(),
+            "vqvae_loss": self.vq_loss_tracker.result(),
+        }
